+++ src/Client/Api/Exception/InvalidWordTypeException.php
@@ -2,16 +2,20 @@
 
 namespace Weglot\Client\Api\Exception;
 
-use Throwable;
-
 /**
  * Class InvalidWordTypeException
  * @package Weglot\Client\Api\Exception
  */
-class InvalidWordTypeException extends \Exception
+class InvalidWordTypeException extends AbstractException
 {
-    public function __construct($message = 'The given WordType is invalid', $code = 400, Throwable $previous = null)
+    /**
+     * InvalidWordTypeException constructor.
+     */
+    public function __construct()
     {
-        parent::__construct($message, $code, $previous);
+        parent::__construct(
+            'The given WordType is invalid.',
+            WeglotCode::PARAMETERS
+        );
     }
 }