+++ src/Client/Endpoint/Languages.php
@@ -18,26 +18,26 @@ class Languages extends Endpoint
     /**
      * @var string
      */
-    protected $iso_639_1;
+    protected $iso639;
 
     /**
      * Languages constructor.
-     * @param string $iso_639_1     ISO 639-1 code to identify language
+     * @param string $iso639    ISO 639-1 code to identify language
      * @param Client $client
      */
-    public function __construct($iso_639_1, Client $client)
+    public function __construct($iso639, Client $client)
     {
-        $this->setIso639($iso_639_1);
+        $this->setIso639($iso639);
         parent::__construct($client);
     }
 
     /**
-     * @param $iso_639_1
+     * @param $iso639
      * @return $this
      */
-    public function setIso639($iso_639_1)
+    public function setIso639($iso639)
     {
-        $this->iso_639_1 = $iso_639_1;
+        $this->iso639 = $iso639;
 
         return $this;
     }
@@ -57,10 +57,10 @@ class Languages extends Endpoint
     public function handle()
     {
         $data = LanguagesFactory::data();
-        if (!isset($data[$this->iso_639_1])) {
+        if (!isset($data[$this->getIso639()])) {
             throw new InvalidLanguageException();
         }
-        $language = $data[$this->iso_639_1];
+        $language = $data[$this->getIso639()];
 
         $factory = new LanguagesFactory($language);
         return $factory->handle();